moved csv button to mapedit.fxml. the button is no longer on its own page. editdatabasecontroller and fxml have been deleted.
