+++ src/main/java/edu/wpi/cs3733/d19/teamO/controller/ControllerModule.java
@@ -91,7 +91,7 @@ public class ControllerModule extends AbstractModule {
     install(new FactoryModuleBuilder().build(SchedulingController.Factory.class));
     install(new FactoryModuleBuilder().build(SchedulingViewController.Factory.class));
     install(new FactoryModuleBuilder().build(SchedulingCalendarController.Factory.class));
-    install(new FactoryModuleBuilder().build(EditNodeDatabaseController.Factory.class));
+//    install(new FactoryModuleBuilder().build(EditNodeDatabaseController.Factory.class));
 
 
     install(new FactoryModuleBuilder().build(AudioVisualController.Factory.class));
