+++ src/main/java/edu/wpi/cs3733/d19/teamO/controller/MapEditController.java
@@ -1,5 +1,10 @@
 package edu.wpi.cs3733.d19.teamO.controller;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.io.Writer;
 import java.util.Optional;
 import java.util.Set;
 
@@ -7,6 +12,8 @@ import com.google.inject.Inject;
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXComboBox;
 import com.jfoenix.controls.JFXTextField;
+import com.opencsv.exceptions.CsvDataTypeMismatchException;
+import com.opencsv.exceptions.CsvRequiredFieldEmptyException;
 
 import javafx.fxml.FXML;
 import javafx.scene.Parent;
@@ -17,10 +24,13 @@ import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.BorderPane;
+import javafx.stage.DirectoryChooser;
 
 import edu.wpi.cs3733.d19.teamO.component.MapView2;
 import edu.wpi.cs3733.d19.teamO.entity.Edge;
 import edu.wpi.cs3733.d19.teamO.entity.Node;
+import edu.wpi.cs3733.d19.teamO.entity.csv.EdgeCsvReaderWriter;
+import edu.wpi.cs3733.d19.teamO.entity.csv.NodeCsvReaderWriter;
 import edu.wpi.cs3733.d19.teamO.entity.database.Database;
 
 @SuppressWarnings("PMD.TooManyFields")
@@ -76,6 +86,8 @@ public class MapEditController implements Controller {
   Tab mapTab;
   @FXML
   Tab tableTab;
+  @FXML
+  private JFXButton exportButton;
 
   @FXML
   TableColumn<Node, String> nodeIdCol;
@@ -100,7 +112,6 @@ public class MapEditController implements Controller {
 
   @FXML
   void initialize() {
-
 
     nodeTypeComboBox.getItems().addAll(Node.NodeType.values());
 
@@ -358,4 +369,32 @@ public class MapEditController implements Controller {
   public interface Factory {
     MapEditController create();
   }
+
+  /**
+   * Export to a csv file.
+   */
+
+  @FXML
+  public void onExportButtonAction() throws IOException, CsvDataTypeMismatchException,
+      CsvRequiredFieldEmptyException {
+    final DirectoryChooser directoryChooser = new DirectoryChooser();
+    final File selectedDirectory = directoryChooser.showDialog(getRoot().getScene().getWindow());
+    final File nodeFile = new File(selectedDirectory, "nodes.csv");
+    final File edgesFile = new File(selectedDirectory, "edges.csv");
+
+    if (selectedDirectory != null) {
+      Writer fw = new OutputStreamWriter(new FileOutputStream(nodeFile), "UTF-8");
+
+      NodeCsvReaderWriter nodeCsvReaderWriter = new NodeCsvReaderWriter();
+      nodeCsvReaderWriter.writeNodes(fw, database.getAllNodes());
+
+      fw = new OutputStreamWriter(new FileOutputStream(edgesFile), "UTF-8");
+      EdgeCsvReaderWriter edgeCsvReaderWriter = new EdgeCsvReaderWriter(database);
+      edgeCsvReaderWriter.writeNodes(fw, database.getAllEdges());
+
+    }
+  }
+
+
+
 }
