+++ src/main/resources/edu/wpi/cs3733/d19/teamO/controller/AdminWindow.fxml
@@ -41,15 +41,7 @@
                 <Font size="45.0" />
               </font>
             </JFXButton>
-            <JFXButton fx:id="editdatabaseButton" onAction="#editdatabaseOnAction" prefHeight="349.0" prefWidth="484.0" ripplerFill="#3b16b2" style="-fx-background-color: #004E86;" stylesheets="@style/KioskStyle.css" text="NODE DATABASE" GridPane.columnIndex="1">
-              <GridPane.margin>
-                <Insets bottom="1.0" left="1.0" right="1.0" top="1.0" />
-              </GridPane.margin>
-              <font>
-                <Font size="40.0" />
-              </font>
-            </JFXButton>
-            <JFXButton fx:id="employeeButton" onAction="#employeeOnAction" prefHeight="368.0" prefWidth="740.0" style="-fx-background-color: #004E86;" stylesheets="@style/KioskStyle.css" text="EMPLOYEE MANAGEMENT" GridPane.columnIndex="1" GridPane.rowIndex="1">
+            <JFXButton fx:id="employeeButton" onAction="#employeeOnAction" prefHeight="368.0" prefWidth="740.0" style="-fx-background-color: #004E86;" stylesheets="@style/KioskStyle.css" text="EMPLOYEE MANAGEMENT" GridPane.columnIndex="1">
               <GridPane.margin>
                 <Insets bottom="1.0" left="1.0" right="1.0" top="1.0" />
               </GridPane.margin>
